requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-admin==1.6.1
celery==5.3.6
redis==4.6.0
playwright==1.40.0
requests==2.31.0
sqlalchemy==2.0.25
faker==23.2.0
python-dotenv==1.0.0
locust==2.20.0